PHOENIX-361 Collect property tax screen

+++ egov/egov-demand/src/main/java/org/egov/demand/dao/EgBillDao.java
@@ -48,7 +48,7 @@ public interface EgBillDao {
 
 	public EgBillType getBillTypeByCode(String code);
 
-	public EgBill findById(Integer id, boolean lock);
+	public EgBill findById(Long id, boolean lock);
 
 	public List<EgBill> findAll();
